+++ CHANGELOG.md
@@ -4,5 +4,6 @@ FEATURES:
 
 * **New Data Source:** `clouddk_locations`
 * **New Data Source:** `clouddk_packages`
+* **New Data Source:** `clouddk_server`
 * **New Data Source:** `clouddk_servers`
 * **New Data Source:** `clouddk_templates`